day 9. yay!! I caught up! the IntCode computer really needs some tlc, very cumbersome to work with at the moment

+++ src/main/kotlin/day2.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.runBlocking
 @ExperimentalCoroutinesApi
 fun main() = runBlocking<Unit> {
     productOf((0..99), (0..99))
-        .first { (noun, verb) -> configureProgram(noun, verb, programInput()).run() == 19690720 }
+        .first { (noun, verb) -> configureProgram(noun, verb, programInput()).run() == 19690720L }
         .let { (noun, verb) ->
             println("Sir... we got him! \uD83D\uDC4D")
             println("noun: $noun verb: $verb answer: ${(100 * noun) + verb}")
@@ -13,14 +13,14 @@ fun main() = runBlocking<Unit> {
 }
 
 @ExperimentalCoroutinesApi
-fun configureProgram(noun: Int, verb: Int, instruction: List<Int>): Program {
+fun configureProgram(noun: Int, verb: Int, instruction: List<Long>): Program {
     val workingMemory = instruction.toMutableList()
-    workingMemory[1] = noun
-    workingMemory[2] = verb
+    workingMemory[1] = noun.toLong()
+    workingMemory[2] = verb.toLong()
     return Program(workingMemory)
 }
 
-fun programInput() : List<Int> =
+fun programInput() : List<Long> =
     readTextFromFile("day2.txt")
         .split(",")
-        .map { it.toInt() }
+        .map { it.toLong() }
